EDITORIAL: Adjust file names for single-feature repository (#47)
Replaces instances of real-world-geometry as part of the repository path with the updated plane-detection.
Removes plane-detection- prefix on several files to match the naming of those files in other single-feature repositories
Removes webxrmeshing-1.bs, which was the precursor to the real-world-meshing repository.
Updates Makefile and pr-preview.json for new index.bs naming and to only worry about that file.